pirogue-admin/pyproject.toml
[build-system]
requires = ["setuptools==66.1.1"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["pirogue_admin*"]
namespaces = false


[project]
name = "pirogue-admin"
version = "2.0.11"
description = "Admin management for the PiRogue"
readme = "README.md"
license = { text = "GPL-3.0-or-later" }

requires-python = ">= 3.9"
dependencies = [
  "grpcio",
  "pirogue_admin_api==2.0.3",
  "psutil~=5.9.4",
  "pyroute2",
  "pystemd~=0.11.0",
  "pyyaml",
  "requests",
  "rich"
]

authors = [
  {name = "Cyril Brulebois", email = "cyril@debamax.com"},
]

maintainers = [
  {name = "U+039b", email = "hello@pts-project.org"}
]

[project.scripts]
pirogue-admin = "pirogue_admin.cmd.cli:main"
pirogue-admin-daemon = "pirogue_admin.daemon:serve"
pirogue-admin-wireguard = "pirogue_admin.cmd.wireguard_cli:main"
network-configuration-monitor = "pirogue_admin.network_configuration_monitor:main"

[project.urls]
repository = "https://github.com/PiRogueToolSuite/pirogue-admin"
issues = "https://github.com/PiRogueToolSuite/pirogue-admin/issues"
